Former-commit-id: 4515878425729344d36c7f68f8b1bc4b7539bc2c

+++ backend/src/main/java/com/sss/tally/api/calculate/dto/CalculateDto.java
@@ -146,7 +146,7 @@ public class CalculateDto {
 
 		private int myAmount;
 
-		private Long allAmount;
+		private int allAmount;
 
 		private LocalDateTime paymentDate;
